+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added:
 - Empty Dashboards instructions
+- Delete dashboards confirmation in red
 
 ## [0.8.29] - 2022-06-01
 ### Added: